refactor(dpapi-fuzzing): small refactoring and code clean-up;
