+++ crates/dpapi-fuzzing/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "dpapi-fuzzing"
 version = "0.1.0"
-edition = "2024"
+edition = "2021"
 
 [dependencies]
 dpapi-pdu = { workspace = true, features = ["arbitrary"] }
