+++ crates/dpapi-fuzzing/src/oracles/mod.rs
@@ -1,0 +1,5 @@
+mod decoding;
+mod round_trip;
+
+pub use decoding::structure_decoding;
+pub use round_trip::round_trip;
