+++ fuzz/fuzz_targets/dpapi_round_trip.rs
@@ -1,10 +1,10 @@
 #![no_main]
 
-extern crate libfuzzer_sys;
 extern crate dpapi_fuzzing;
+extern crate libfuzzer_sys;
 
 use dpapi_fuzzing::generator::AnyStruct;
-use dpapi_fuzzing::round_trip;
+use dpapi_fuzzing::oracles::round_trip;
 
 libfuzzer_sys::fuzz_target!(|any: AnyStruct| {
     round_trip(any);
