+++ fuzz/fuzz_targets/fuzz_ts_request.rs
@@ -1,5 +1,6 @@
 #![no_main]
-#[macro_use] extern crate libfuzzer_sys;
+#[macro_use]
+extern crate libfuzzer_sys;
 extern crate sspi;
 
 use sspi::credssp::TsRequest;
@@ -9,6 +10,6 @@ fuzz_target!(|data: &[u8]| {
         let _req_len = req.buffer_len();
         let _result = req.check_error();
     }
-    
+
     let _creds = sspi::credssp::read_ts_credentials(data);
 });
